requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
django-debug-toolbar
Pillow
psycopg2
Fabric3